refactor leo core modules
